update argo compiler to support pod metadata
